+++ packages/ariakit/src/select/__examples__/select/test.tsx
@@ -143,7 +143,7 @@ test("navigate through items with keyboard", async () => {
   expect(getOption("Apple")).toHaveFocus();
 });
 
-test("typeahead visible", async () => {
+test("typeahead open", async () => {
   render(<Example />);
   await press.Tab();
   await press.Enter();